tests: update OnspringRequest tests
